requirements.txt
requests
easyocr
opencv-python
numpy
Pillow
keyboard
pynput
pyautogui
fuzzywuzzy
python-Levenshtein